#42 Resolved some TODOs
