+++ src/main/java/ee/yals/controllers/rest/StoreRestController.java
@@ -6,13 +6,11 @@ import ee.yals.json.ErrorJson;
 import ee.yals.json.StoreRequestJson;
 import ee.yals.json.StoreResponseJson;
 import ee.yals.json.internal.Json;
-import ee.yals.models.Token;
 import ee.yals.models.User;
-import ee.yals.models.dao.TokenDao;
-import ee.yals.models.dao.UserDao;
 import ee.yals.result.GetResult;
 import ee.yals.result.StoreResult;
 import ee.yals.services.LinkService;
+import ee.yals.services.token.TokenService;
 import ee.yals.utils.AppUtils;
 import ee.yals.utils.UrlExtraValidator;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,8 +25,9 @@ import javax.validation.ConstraintViolation;
 import javax.validation.Validation;
 import javax.validation.Validator;
 import java.util.HashSet;
-import java.util.Optional;
 import java.util.Set;
+
+import static ee.yals.services.token.DbTokenService.NO_TOKEN;
 
 /**
  * Stores long link to storage
@@ -38,20 +37,12 @@ import java.util.Set;
 @RestController
 public class StoreRestController {
 
-    private static final String NO_TOKEN = "NOT_A_TOKEN";
-
     @Autowired
     @Qualifier("dbStorage")
     private LinkService linkService;
 
-    //TODO call service instead direct access from Controller to DAO
     @Autowired
-    private UserDao userDao;
-
-    //TODO call service instead direct access from Controller to DAO
-    @Autowired
-    private TokenDao tokenDao;
-
+    private TokenService tokenService;
 
     @RequestMapping(method = {RequestMethod.POST, RequestMethod.PUT},
             value = Endpoint.STORE_API)
@@ -96,7 +87,7 @@ public class StoreRestController {
         }
 
         //TODO get token from JSON
-        User linkOwner = giveUserFromToken(NO_TOKEN);
+        User linkOwner = tokenService.giveUserFromToken(NO_TOKEN);
 
         StoreResult result = linkService.storeNew(ident, storeInput.getLink(), linkOwner);
         if (result instanceof StoreResult.Success) {
@@ -117,19 +108,4 @@ public class StoreRestController {
         return (searchResult instanceof GetResult.Success);
     }
 
-    private User giveUserFromToken(String token) {
-        //TODO call services here
-        if (token.equals(NO_TOKEN)) {
-            //default user
-            return userDao.getDefaultUser();
-        } else {
-            //TODO call services here
-            Optional<Token> tokenRecord = tokenDao.findSingleByToken(token);
-            if (tokenRecord.isPresent()) {
-                return tokenRecord.get().getOwner();
-            } else {
-                return userDao.getDefaultUser();
-            }
-        }
-    }
 }
