+++ src/main/java/ee/yals/services/token/TokenService.java
@@ -1,0 +1,13 @@
+package ee.yals.services.token;
+
+import ee.yals.models.User;
+
+/**
+ * Class description
+ *
+ * @since 3.0
+ */
+public interface TokenService {
+
+    User giveUserFromToken(String token);
+}
