+++ src/main/java/ee/yals/services/user/UserService.java
@@ -1,0 +1,12 @@
+package ee.yals.services.user;
+
+import ee.yals.models.User;
+
+/**
+ * Service which queries from DB about users
+ *
+ * @since 3.0
+ */
+public interface UserService {
+    User getDefaultUser();
+}
